standardization: Create a python package of zircbot
Currently zircbot is a python script. It does one thing and does it
well, transform Trello event into IRC message.

We can imagine that Trello is just a backend and that we might want to
implement others with the notion of plugins/drivers.

In order to add more feature to this module applying some python
packaging rules from the beginning might be a good idea